Refactor the Memory Module to be more robust and more stateless
~2hrs
- The IMemoryManager::deallocate function now requires the user to pass in the allocated size. This helps makes some implmentation details simpler.
- Updated any code that to follow this new standard.
- 'bf::path::append' now returns if the path is truncated.
- Fixed a bug in the Job system where the parent argument was ignored for one of the overloads od the makeTask functions.
- Stripped down the old C Array API to just the stuff that is currently beign used.
- More Skeletal animation work.

+++ BifrostScript/src/bifrost_vm_parser.c
@@ -354,8 +354,8 @@ static inline bfBool32 bfParser_is(BifrostParser* self, bfTokenType type)
 
 static inline GrammarRule typeToRule(bfTokenType type)
 {
-  assert((size_t)type < bfCArraySize(s_Rules) && "Invalid token type.");
-  return s_Rules[type];
+  assert(((size_t)type < bfCArraySize(s_Rules)) && "Invalid token type.");
+  return s_Rules[(size_t)type];
 }
 
 #define builder() self->fn_builder
